"moving routes into controllers"
