+++ controllers/auth-routes.js
@@ -1,10 +1,10 @@
 // Create user   /api/users
 const router = require('express').Router();
-const User = require('../../models/user');
+const User = require('../models/user');
 const bcrypt = require('bcryptjs');
 const { Op } = require('sequelize');
 
-//localhost:3001/api/users/
+//localhost:3001/auth/
 
 router.post('/signup', async (req, res) => {
     const {
@@ -14,11 +14,14 @@ router.post('/signup', async (req, res) => {
         password,
         reEnterPassword,
     } = req.body;
+
     //might need to add additional parameters for reEnterPassword
     if (!username || !email || !password) {
         //return res.status(400).json({ error: 'You must provider username, first name, last name, password, and email' });
-        res.redirect('../../signup');
+        res.redirect('../signup');
     }
+
+	console.log("Sign Up Route 2");
 
     // PaSsWoRd
     try {
@@ -35,11 +38,13 @@ router.post('/signup', async (req, res) => {
         //set session data
         CreateSessionData(req, newUser);
 
-        res.redirect('../../dashboard');
+		console.log("success");
+
+        res.redirect('../results');
     } catch (e) {
         console.log(e);
         // res.json(e);
-        res.redirect('../../signup');
+        res.redirect('../signup');
     }
 
 });
@@ -52,7 +57,7 @@ router.post('/login', async (req, res) => {
     const { username, password } = req.body;
     if (!username || !password) {
         //return res.status(401).json({ error: 'You must provide a valid username and password' });
-        res.redirect('../../');
+        res.redirect('../');
     }
     try {
         //	find the user with the given email
@@ -65,22 +70,22 @@ router.post('/login', async (req, res) => {
         if (user == null) {
             //	if no user exists, give them a 400
             //return res.status(400).json({ error: 'No user with that username' });
-            res.redirect('../../');
+            res.redirect('../');
         }
         //	take the users hashed password and compare it with the password that they're passing in from the form
         const isMatching = await bcrypt.compare(password, user.password)
         console.log(isMatching);
         if (!isMatching) {
             //return res.status(401).json({ error: 'Invalid password' });
-            res.redirect('../../');
+            res.redirect('../');
         }
         //return res.status(200).json({ message: 'You are now logged in successfully' });
         CreateSessionData(req, user);
 
-        res.redirect('../../dashboard');
+        res.redirect('../results');
     } catch (e) {
         console.log(e);
-        res.redirect('../../');
+        res.redirect('../');
         //return res.status(200).json(e);
     }
 });
