--- routes/api/bookRoutes/index.js
@@ -1,6 +1,0 @@
-const router = require('express').Router();
-const userRoutes = require('./userRoutes');
-
-router.use('/users', userRoutes);
-
-module.exports = router;
