--- routes/api/index.js
@@ -1,8 +1,0 @@
-const router = require('express').Router();
-const userRoutes = require('./userRoutes');
-
-//localhost
-router.use('/users', userRoutes);
-//router.use("/books", bookRoutes)
-
-module.exports = router;
