+++ views/signup.handlebars
@@ -20,7 +20,7 @@
             bg-orange-100 bg-opacity-60 backdrop-blur-md
             shadow-white shadow-inner drop-shadow-2xl rounded-lg" >
             
-            <form class = "w-full h-3/4 flex flex-col justify-evenly items-center text-orange-900 " autocomplete="off" >
+            <form  method="post" action="auth/signup" class = "w-full h-3/4 flex flex-col justify-evenly items-center text-orange-900 " autocomplete="off" >
 
             <h3 class = "text-orange-950 text-6xl text-center">Welcome to the Club Bookworm!...</h3>
